feat: split config and copy docs files to `zh` subdirectory move English specific config to en.ts and create the zh.ts for Chinese Simplified translate the config string
