+++ docs/.vitepress/en.ts
@@ -1,0 +1,101 @@
+import type { DefaultTheme } from 'vitepress'
+import { defineConfig } from 'vitepress'
+import { version } from '../../package.json'
+
+const GUIDES: DefaultTheme.NavItemWithLink[] = [
+  { text: 'Getting Started', link: '/guide/' },
+  { text: 'Installation', link: '/guide/install' },
+  { text: 'Bundles', link: '/guide/bundles' },
+  { text: 'Dual Themes', link: '/guide/dual-themes' },
+  { text: 'Transformers', link: '/guide/transformers' },
+  { text: 'Theme Colors Manipulation', link: '/guide/theme-colors' },
+  { text: 'Compatibility Build', link: '/guide/compat' },
+  { text: 'Custom Themes', link: '/guide/load-theme' },
+  { text: 'Custom Languages', link: '/guide/load-lang' },
+]
+
+const REFERENCES: DefaultTheme.NavItemWithLink[] = [
+  { text: 'Themes', link: '/themes' },
+  { text: 'Languages', link: '/languages' },
+]
+
+const INTEGRATIONS: DefaultTheme.NavItemWithLink[] = [
+  { text: 'TypeScript Twoslash', link: '/packages/twoslash' },
+  { text: 'Markdown It', link: '/packages/markdown-it' },
+  { text: 'Rehype', link: '/packages/rehype' },
+  { text: 'Monaco Editor', link: '/packages/monaco' },
+  { text: 'VitePress', link: '/packages/vitepress' },
+  { text: 'Astro', link: '/packages/astro' },
+  { text: 'Common Transformers', link: '/packages/transformers' },
+  { text: 'CLI', link: '/packages/cli' },
+]
+
+const VERSIONS: DefaultTheme.NavItemWithLink[] = [
+  { text: `v${version} (current)`, link: '/' },
+  { text: `Release Notes`, link: 'https://github.com/antfu/shikiji/releases' },
+  { text: `Contributing`, link: 'https://github.com/antfu/shikiji/blob/main/CONTRIBUTING.md' },
+]
+
+export const enConfig = defineConfig({
+  lang: 'en-US',
+  description: 'A beautiful and powerful syntax highlighter',
+
+  themeConfig: {
+    nav: [
+      {
+        text: 'Guide',
+        items: [
+          {
+            items: GUIDES,
+          },
+        ],
+      },
+      {
+        text: 'Integrations',
+        items: INTEGRATIONS,
+      },
+      {
+        text: 'References',
+        items: REFERENCES,
+      },
+      // {
+      //   text: 'Play',
+      //   link: '/play',
+      // },
+      {
+        text: `v${version}`,
+        items: VERSIONS,
+      },
+    ],
+
+    sidebar: Object.assign(
+      {},
+      {
+        '/': [
+          {
+            text: 'Guide',
+            items: GUIDES,
+          },
+          {
+            text: 'Integrations',
+            items: INTEGRATIONS,
+          },
+          {
+            text: 'References',
+            items: REFERENCES,
+          },
+        ],
+      },
+    ),
+
+    editLink: {
+      pattern: 'https://github.com/antfu/shikiji/edit/main/docs/:path',
+      text: 'Suggest changes to this page',
+    },
+
+    footer: {
+      message: 'Released under the MIT License.',
+      copyright: 'Copyright © 2021 Pine Wu, 2023-PRESENT Anthony Fu.',
+    },
+  },
+})
