+++ docs/zh/index.md
@@ -1,0 +1,34 @@
+---
+layout: home
+
+hero:
+  name: "式辞 Shikiji"
+  text: "Syntax highlighter"
+  tagline: A beautiful and powerful syntax highlighter
+  image:
+    src: /logo.svg
+    alt: Shikiji Logo
+  actions:
+    - theme: brand
+      text: Get Started
+      link: /guide/
+    - theme: alt
+      text: Installation
+      link: /guide/install/
+
+features:
+  - title: Accurate & Beautiful
+    icon: 🌈
+    details: TextMate grammar powered, as accurate as your Text Editor. Improves as your Text Editor improves.
+  - title: Zero Runtime
+    icon: ⏱️
+    details: Runs ahead of time, ship zero JavaScript and get the perfect syntax highlight.
+  - title: Customizable
+    icon: 🧩
+    details: HAST based. Addons, transformations, highly customizable.
+  - title: ESM & Universal
+    icon: 🎄
+    details: Fully tree-shakable ESM, runs on any JavaScript runtime, including Browser, Node.js, Cloudflare Workers, etc.
+---
+
+<HomeDemo />
