+++ docs/zh/languages.md
@@ -1,0 +1,9 @@
+# Languages
+
+Language grammars listed below are re-distributed via [`tm-grammars`](https://github.com/antfu/textmate-grammars-themes/tree/main/packages/tm-grammars) into the `shikiji` package.
+
+<LanguagesList />
+
+Grammars are covered by their repositories’ respective licenses, which are permissive (apache-2.0, mit, etc), and made available in [this NOTICE](https://github.com/antfu/textmate-grammars-themes/blob/main/packages/tm-grammars/NOTICE).
+
+For loading your custom languages, please reference to [this guide](/guide/load-lang).
