+++ docs/zh/packages/astro.md
@@ -1,0 +1,5 @@
+# Astro Integration
+
+You don't need explicit integration in [Astro](https://astro.build/) as Astro already uses Shikiji under the hood.
+
+Learn more about the configurations in the [Astro documentation](https://docs.astro.build/en/guides/markdown-content/#syntax-highlighting).
