+++ docs/zh/play.md
@@ -1,0 +1,5 @@
+---
+layout: home
+---
+
+<ShikijiPlayground />
