+++ docs/zh/themes.md
@@ -1,0 +1,9 @@
+# Themes
+
+Themes listed below are re-distributed via [`tm-themes`](https://github.com/antfu/textmate-grammars-themes/tree/main/packages/tm-themes) into the `shikiji` package.
+
+<ThemesList />
+
+Themes are covered by their repositories’ respective licenses, which are permissive (apache-2.0, mit, etc), and made available in [this NOTICE](https://github.com/antfu/textmate-grammars-themes/blob/main/packages/tm-themes/NOTICE).
+
+For loading your custom themes, please reference to [this guide](/guide/load-theme).
